Fixed number cast when post empty string (#649)
* Fixed number cast when post empty string

* feat: Remove deprecated DoubleCast and LongCast classes

The DoubleCast and LongCast classes have been deprecated since v3.1. They have been replaced by FloatCast and IntegerCast respectively. This commit removes the deprecated classes to clean up the codebase and prepare for their removal in v3.2.

* feat: Update LongCast to extend IntegerCast

The LongCast class has been updated to extend the IntegerCast class, deprecating the previous implementation. This change improves code organization and prepares for the removal of LongCast in v3.2.

---------

Co-authored-by: Deeka Wong <8337659+huangdijia@users.noreply.github.com>