feat: Add default name parameter to Cache constructor
The recent code changes added a default value for the `name` parameter in the `Cache` constructor. This allows the `name` parameter to be optional when creating a new `Cache` instance.

Based on the recent user commits and repository commits, the commit message follows the established convention of using a prefix to indicate the type of change (`feat` for a new feature).