Use `krowinski/php-mysql-replication` instead of `moln/php-mysql-replication` (#706)
* Use `krowinski/php-mysql-replication` instead of `moln/php-mysql-replication`

* feat: Reconnect to MySQL replication on connection loss

The code changes in `Consumer.php` handle the scenario where a `SocketException` is caught, indicating a lost connection to the MySQL replication. Instead of using the `retry` function, the code now includes a placeholder comment to implement the reconnection logic. This change improves the resilience of the application by ensuring that the connection is reestablished in case of a failure.

* json fix

* Adapter 8.x

* fix phpstan

* chore: Update php-mysql-replication dependency to version 8.0

* feat: Remove src/trigger for PHP 8.1

The code changes in `.github/workflows/tests.yaml` remove the `src/trigger` directory when the PHP version is 8.1. This change is made to address a specific compatibility issue with PHP 8.1.

Co-authored-by: Deeka Wong <8337659+huangdijia@users.noreply.github.com>

* chore: Update php-mysql-replication dependency to version 7.0 or 8.0

* feat: Update php-mysql-replication dependency to version 7.0 or 8.0

* chore: Update dependencies and remove unnecessary code

The code changes in `.github/workflows/tests.yaml` update the dependencies using `composer update -o` and remove the `src/trigger` directory when the PHP version is 8.1. These changes improve the compatibility and performance of the application.

Co-authored-by: Deeka Wong <8337659+huangdijia@users.noreply.github.com>

* chore: Update phpstan.neon.dist

The code changes in `phpstan.neon.dist` remove the commented out line for the `src/trigger` directory exclusion. This change improves the configuration file by removing unnecessary code.

Co-authored-by: Deeka Wong <8337659+huangdijia@users.noreply.github.com>

* chore: Update HealthMonitor and SnapshotSubscriber to handle nullable BinLogCurrent

The code changes in HealthMonitor.php and SnapshotSubscriber.php modify the setBinLogCurrent method to accept a nullable BinLogCurrent object. This change ensures compatibility with scenarios where the BinLogCurrent object may be null. It improves the robustness and flexibility of the application.

Co-authored-by: Deeka Wong <8337659+huangdijia@users.noreply.github.com>

* feat: Update TriggerSubscriber to handle nullable BinLogCurrent

The code changes in TriggerSubscriber.php modify the values assignment to handle nullable BinLogCurrent objects. This change ensures compatibility with scenarios where the BinLogCurrent object may be null. It improves the robustness and flexibility of the application.

* feat: Update setBinLogCurrent method to accept non-nullable BinLogCurrent object

---------

Co-authored-by: Deeka Wong <8337659+huangdijia@users.noreply.github.com>